Made quick changes to CharSpawner and Note

+++ src/Note.java
@@ -8,7 +8,7 @@ public class Note {
 	//font size would be the size variable
 	private String noteType;
 	//represents letter/number of object
-	public Note(int x, int y) {
+	public Note(double x, double y) {
 		posX = x;
 		posY = y;
 	}
